requirements.txt
dj-database-url==0.3.0
Django==1.8.5
django-cors-headers==1.1.0
django-rest-swagger==0.3.4
djangorestframework==3.2.4
drf-nested-routers==0.10.0
fake-factory==0.5.3
gunicorn==19.3.0
psycopg2==2.6.1
PyYAML==3.11
wheel==0.24.0
whitenoise==2.0.4
